pyproject.toml
[tool.poetry]
name = "dashboard"
version = "0.1.0"
description = ""
authors = ["rede akbar wijaya <ribonred@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
django-extensions = "^3.2.1"
django-simple-history = "^3.2.0"
django-structlog = "^4.0.1"
django-unfold = "^0.7.0"
django-debug-toolbar = "^3.8.1"
django-lifecycle = "^1.0.0"
daphne = "^4.0.0"
djangorestframework = "^3.14.0"
drf-spectacular = "0.26.1"
django-filter = "^23.2"
faker = "^15.3.4"

[tool.poetry.group.linting.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.7.0"
pre-commit = "^2.21.0"

[tool.poetry.group.remote.dependencies]
gunicorn = "^20.1.0"
uvicorn = {extras = ["standart"], version = "^0.21.1"}
whitenoise = "^6.5.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.0"
pytest-sugar = "^0.9.6"
pytest-mock = "^3.10.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
